improve config saving clean arch
