+++ src/domain/usecases/SaveConfigUseCase.ts
@@ -1,44 +1,33 @@
-import { isEmpty, isEqual } from "lodash";
-
 import { UseCase } from "../../webapp/CompositionRoot";
 import { ConfigRepository } from "../repositories/ConfigRepository";
-import { CustomText, getDefaultCustomText } from "../entities/CustomText";
-import { Config, PartialConfig } from "../entities/Config";
+import { CustomText } from "../entities/CustomText";
+import { Config } from "../entities/Config";
+import { Permission } from "../entities/Permission";
+
+export type PartialConfig = Pick<Partial<Config>, "showAllModules" | "logo"> & {
+    settingsPermissions?: Partial<Permission>;
+    customText?: Partial<CustomText>;
+};
 
 export class SaveConfigUseCase implements UseCase {
     constructor(private configRepository: ConfigRepository) {}
 
-    public async execute(config: PartialConfig): Promise<Config> {
-        const { customText, ...rest } = config;
+    public async execute(update: PartialConfig): Promise<void> {
+        const config = await this.configRepository.get();
 
-        const configUpdates = {
-            ...rest,
-            ...(customText && { customText: this.cleanCustomText(customText) }),
+        const updatedConfig: Config = {
+            settingsPermissions: {
+                users: update.settingsPermissions?.users ?? config.settingsPermissions?.users ?? [],
+                userGroups: update.settingsPermissions?.userGroups ?? config.settingsPermissions?.userGroups ?? [],
+            },
+            showAllModules: update.showAllModules ?? config.showAllModules,
+            logo: update.logo ?? config.logo,
+            customText: {
+                rootTitle: update.customText?.rootTitle ?? config.customText.rootTitle,
+                rootSubtitle: update.customText?.rootSubtitle ?? config.customText.rootSubtitle,
+            },
         };
 
-        return await this.configRepository.save(configUpdates);
-    }
-
-    private cleanCustomText(customText: Partial<CustomText>): PartialConfig["customText"] {
-        const defaultCustomText = getDefaultCustomText();
-
-        return Object.entries(customText).reduce((acc, [key, value]) => {
-            if (isEmpty(value)) return acc;
-
-            const defaultValue = defaultCustomText[key as keyof CustomText];
-            const hasEqualReferenceValue = value.referenceValue === defaultValue?.referenceValue;
-            const hasEmptyOrEqualTranslation =
-                isEmpty(value.translations) ||
-                Object.values(value.translations || {}).every(
-                    translation => !translation || translation.trim() === ""
-                ) ||
-                isEqual(value.translations, defaultValue?.translations);
-
-            if (hasEqualReferenceValue && hasEmptyOrEqualTranslation) {
-                return { ...acc, [key]: undefined };
-            }
-
-            return { ...acc, [key]: value };
-        }, {});
+        return await this.configRepository.save(updatedConfig);
     }
 }
